OCT-111: moved GetCategoryChildrenQuery
